Move "TypeScript Runtime" configuration to first wizard page. Move "Working sets" configuration to the end of the wizard page. Remove TypeScriptRuntimeWizardPage.
