+++ eclipse/ts.eclipse.ide.ui/src/ts/eclipse/ide/internal/ui/TypeScriptUIMessages.properties
@@ -157,6 +157,14 @@ NewTypeScriptProjectWizard_windowTitle=New TypeScript Project
 NewTypeScriptProjectWizard_newProjectTitle=Create a TypeScript Project
 NewTypeScriptProjectWizard_newProjectDescription=Create a TypeScript project in the workspace or in an external location.
 
+AbstractWizardNewTypeScriptProjectCreationPage_nodejs_group_label=Node.js
+AbstractWizardNewTypeScriptProjectCreationPage_useEmbeddedNodeJs_label=Use embedded
+AbstractWizardNewTypeScriptProjectCreationPage_useInstalledNodeJs_label=Use installed
+
+WizardNewTypeScriptProjectCreationPage_typescript_group_label=TypeScript Runtime
+WizardNewTypeScriptProjectCreationPage_useEmbeddedTsRuntime_label=Use embedded
+WizardNewTypeScriptProjectCreationPage_useInstallTsRuntime_label=Install
+
 TSConfigWizardPage_title=Configure tsconfig.json
 TSConfigWizardPage_description=Configure the tsconfig.json.
 TSConfigWizardPage_module=Module:
@@ -172,15 +180,6 @@ TSConfigWizardPage_noFallthroughCasesInSwitch=No fallthrough in switch
 TSConfigWizardPage_noImplicitAny=No implicit any
 TSConfigWizardPage_noImplicitReturns=No implicit returns
 TSConfigWizardPage_strictNullChecks=Strict null checks
-
-TypeScriptRuntimeAndNodejsWizardPage_title=TypeScript Runtime / Node.js
-TypeScriptRuntimeAndNodejsWizardPage_description=Select the TypeScript runtime and the Node.js required by the TypeScript tsserver.
-TypeScriptRuntimeAndNodejsWizardPage_nodejs_group_label=Node.js
-TypeScriptRuntimeAndNodejsWizardPage_typescript_group_label=TypeScript Runtime
-TypeScriptRuntimeAndNodejsWizardPage_useEmbeddedTsRuntime_label=Use embedded
-TypeScriptRuntimeAndNodejsWizardPage_useInstallTsRuntime_label=Install
-TypeScriptRuntimeAndNodejsWizardPage_useEmbeddedNodeJs_label=Use embedded
-TypeScriptRuntimeAndNodejsWizardPage_useInstalledNodeJs_label=Use installed
 
 # NPM Module widget
 NPMModuleVersionsSelectionDialog_title=Open NPM module version for ''{0}''
